Updates for font, and fixed-width header in subtitle
In poole.css, the html default fonts were set to match the lanyon.css
fons; different sets meant that the about/posts were different than
the landing page. In the site subtitle and the lanyon.css page, a
separate tag was added to make the [digital] part fixed-width.

Font access was unified in the head.html includes file.